refactor: abstract ribbon style
