+++ components/theme/util/genComponentStyleHook.ts
@@ -1,8 +1,9 @@
 /* eslint-disable no-redeclare */
+import { useContext } from 'react';
 import type { CSSInterpolation } from '@ant-design/cssinjs';
 import { useStyleRegister } from '@ant-design/cssinjs';
 import { warning } from 'rc-util';
-import { useContext } from 'react';
+
 import { ConfigContext } from '../../config-provider/context';
 import { genCommonStyle, genLinkStyle } from '../../style';
 import type {
@@ -48,9 +49,14 @@ export type FullToken<ComponentName extends OverrideComponent> = TokenWithCommon
   GlobalTokenWithComponent<ComponentName>
 >;
 
+export type GenStyleFn<ComponentName extends OverrideComponent> = (
+  token: FullToken<ComponentName>,
+  info: StyleInfo<ComponentName>,
+) => CSSInterpolation;
+
 export default function genComponentStyleHook<ComponentName extends OverrideComponent>(
-  component: ComponentName,
-  styleFn: (token: FullToken<ComponentName>, info: StyleInfo<ComponentName>) => CSSInterpolation,
+  componentName: ComponentName | [ComponentName, string],
+  styleFn: GenStyleFn<ComponentName>,
   getDefaultToken?:
     | OverrideTokenWithoutDerivative[ComponentName]
     | ((token: GlobalToken) => OverrideTokenWithoutDerivative[ComponentName]),
@@ -64,6 +70,14 @@ export default function genComponentStyleHook<ComponentName extends OverrideComp
     clientOnly?: boolean;
   },
 ) {
+  const cells = (Array.isArray(componentName) ? componentName : [componentName, componentName]) as [
+    ComponentName,
+    string,
+  ];
+
+  const [component] = cells;
+  const concatComponent = cells.join('-');
+
   return (prefixCls: string): UseComponentStyleResult => {
     const [theme, token, hashId] = useToken();
     const { getPrefixCls, iconPrefixCls, csp } = useContext(ConfigContext);
@@ -93,62 +107,65 @@ export default function genComponentStyleHook<ComponentName extends OverrideComp
     );
 
     return [
-      useStyleRegister({ ...sharedConfig, path: [component, prefixCls, iconPrefixCls] }, () => {
-        const { token: proxyToken, flush } = statisticToken(token);
+      useStyleRegister(
+        { ...sharedConfig, path: [concatComponent, prefixCls, iconPrefixCls] },
+        () => {
+          const { token: proxyToken, flush } = statisticToken(token);
 
-        const customComponentToken = { ...(token[component] as ComponentToken<ComponentName>) };
-        if (options?.deprecatedTokens) {
-          const { deprecatedTokens } = options;
-          deprecatedTokens.forEach(([oldTokenKey, newTokenKey]) => {
-            if (process.env.NODE_ENV !== 'production') {
-              warning(
-                !customComponentToken?.[oldTokenKey],
-                `The token '${String(oldTokenKey)}' of ${component} had deprecated, use '${String(
-                  newTokenKey,
-                )}' instead.`,
-              );
-            }
+          const customComponentToken = { ...(token[component] as ComponentToken<ComponentName>) };
+          if (options?.deprecatedTokens) {
+            const { deprecatedTokens } = options;
+            deprecatedTokens.forEach(([oldTokenKey, newTokenKey]) => {
+              if (process.env.NODE_ENV !== 'production') {
+                warning(
+                  !customComponentToken?.[oldTokenKey],
+                  `The token '${String(oldTokenKey)}' of ${component} had deprecated, use '${String(
+                    newTokenKey,
+                  )}' instead.`,
+                );
+              }
 
-            // Should wrap with `if` clause, or there will be `undefined` in object.
-            if (customComponentToken?.[oldTokenKey] || customComponentToken?.[newTokenKey]) {
-              customComponentToken[newTokenKey] ??= customComponentToken?.[oldTokenKey];
-            }
-          });
-        }
-        const defaultComponentToken =
-          typeof getDefaultToken === 'function'
-            ? getDefaultToken(mergeToken(proxyToken, customComponentToken ?? {}))
-            : getDefaultToken;
+              // Should wrap with `if` clause, or there will be `undefined` in object.
+              if (customComponentToken?.[oldTokenKey] || customComponentToken?.[newTokenKey]) {
+                customComponentToken[newTokenKey] ??= customComponentToken?.[oldTokenKey];
+              }
+            });
+          }
+          const defaultComponentToken =
+            typeof getDefaultToken === 'function'
+              ? getDefaultToken(mergeToken(proxyToken, customComponentToken ?? {}))
+              : getDefaultToken;
 
-        const mergedComponentToken = { ...defaultComponentToken, ...customComponentToken };
+          const mergedComponentToken = { ...defaultComponentToken, ...customComponentToken };
 
-        const componentCls = `.${prefixCls}`;
-        const mergedToken = mergeToken<
-          TokenWithCommonCls<GlobalTokenWithComponent<OverrideComponent>>
-        >(
-          proxyToken,
-          {
-            componentCls,
+          const componentCls = `.${prefixCls}`;
+          const mergedToken = mergeToken<
+            TokenWithCommonCls<GlobalTokenWithComponent<OverrideComponent>>
+          >(
+            proxyToken,
+            {
+              componentCls,
+              prefixCls,
+              iconCls: `.${iconPrefixCls}`,
+              antCls: `.${rootPrefixCls}`,
+            },
+            mergedComponentToken,
+          );
+
+          const styleInterpolation = styleFn(mergedToken as unknown as FullToken<ComponentName>, {
+            hashId,
             prefixCls,
-            iconCls: `.${iconPrefixCls}`,
-            antCls: `.${rootPrefixCls}`,
-          },
-          mergedComponentToken,
-        );
-
-        const styleInterpolation = styleFn(mergedToken as unknown as FullToken<ComponentName>, {
-          hashId,
-          prefixCls,
-          rootPrefixCls,
-          iconPrefixCls,
-          overrideComponentToken: customComponentToken as any,
-        });
-        flush(component, mergedComponentToken);
-        return [
-          options?.resetStyle === false ? null : genCommonStyle(token, prefixCls),
-          styleInterpolation,
-        ];
-      }),
+            rootPrefixCls,
+            iconPrefixCls,
+            overrideComponentToken: customComponentToken as any,
+          });
+          flush(component, mergedComponentToken);
+          return [
+            options?.resetStyle === false ? null : genCommonStyle(token, prefixCls),
+            styleInterpolation,
+          ];
+        },
+      ),
       hashId,
     ];
   };
